Slår sammen splittet brev, refactoring og kombinering av repiterende innholds-blokker.
