+++ pensjonsmaler/src/test/kotlin/no/nav/pensjon/brev/Fixtures.kt
@@ -124,8 +124,7 @@ object Fixtures {
             VedtakOmFjerningAvOmsorgsopptjeningDto::class -> createVedtakOmFjerningAvOmsorgsopptjeningDto() as T
             Gjenlevenderett2027Dto::class -> createGjenlevenderett2027Dto() as T
             EndringAvAlderspensjonSivilstandDto::class -> createEndringAvAlderspensjonSivilstandDto() as T
-            AvslagForLiteTrygdetidAP2011Dto::class -> createAvslagForLiteTrygdetidAP2011Dto() as T
-            AvslagForLiteTrygdetidAP2016Dto::class -> createAvslagForLiteTrygdetidAP2016Dto() as T
+            AvslagForLiteTrygdetidAPDto::class -> createAvslagForLiteTrygdetidAPDto() as T
 
 
             else -> throw IllegalArgumentException("Don't know how to construct: ${letterDataType.qualifiedName}")
