+++ pensjonsmaler/src/test/kotlin/no/nav/pensjon/brev/maler/redigerbar/AvslagForLiteTrygdetidAPTest.kt
@@ -5,32 +5,32 @@ import no.nav.brev.brevbaker.TestTags
 import no.nav.brev.brevbaker.renderTestHtml
 import no.nav.brev.brevbaker.renderTestPDF
 import no.nav.pensjon.brev.Fixtures
-import no.nav.pensjon.brev.api.model.maler.redigerbar.AvslagForLiteTrygdetidAP2011Dto
+import no.nav.pensjon.brev.api.model.maler.redigerbar.AvslagForLiteTrygdetidAPDto
 import no.nav.pensjon.brev.template.Language
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 
 @Tag(TestTags.MANUAL_TEST)
-class AvslagForLiteTrygdetidAP2016Test {
+class AvslagForLiteTrygdetidAPTest {
 
     @Test
     fun testPdf() {
         LetterTestImpl(
-            AvslagForLiteTrygdetidAP2016.template,
-            Fixtures.create<AvslagForLiteTrygdetidAP2011Dto>(),
-            Language.Bokmal,
+            AvslagForLiteTrygdetidAP.template,
+            Fixtures.create<AvslagForLiteTrygdetidAPDto>(),
+            Language.English,
             Fixtures.felles
-        ).renderTestPDF(AvslagForLiteTrygdetidAP2016.kode.name)
+        ).renderTestPDF(AvslagForLiteTrygdetidAP.kode.name)
     }
 
     @Test
     fun testHtml() {
         LetterTestImpl(
-            AvslagForLiteTrygdetidAP2016.template,
-            Fixtures.create<AvslagForLiteTrygdetidAP2011Dto>(),
+            AvslagForLiteTrygdetidAP.template,
+            Fixtures.create<AvslagForLiteTrygdetidAPDto>(),
             Language.Bokmal,
             Fixtures.felles
-        ).renderTestHtml(AvslagForLiteTrygdetidAP2016.kode.name)
+        ).renderTestHtml(AvslagForLiteTrygdetidAP.kode.name)
     }
 }
